Cleaned oneOf package
