--- src/common/models/billingmodels/ratedusageoneof/serverdetails.go
@@ -1,15 +1,0 @@
-package ratedusageoneof
-
-import "github.com/phoenixnap/go-sdk-bmc/billingapi"
-
-type ServerDetails struct {
-	Id       string `json:"id" yaml:"id"`
-	Hostname string `json:"hostname" yaml:"hostname"`
-}
-
-func ServerDetailsFromSdk(serverDetails *billingapi.ServerDetails) *ServerDetails {
-	return &ServerDetails{
-		Id:       serverDetails.Id,
-		Hostname: serverDetails.Hostname,
-	}
-}
